# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: LLT Assistant API
  description: |
    Comprehensive Low-Level Test (LLT) Assistant API for VSCode extension.
    
    **Core Features:**
    1. **Test Generation** - Generate pytest test cases from minimal user input
    2. **Coverage Optimization** - Analyze and improve test coverage
    3. **Dynamic Maintenance** - Detect code changes and update tests accordingly
    4. **Quality Analysis** - Detect defects and redundancies in existing tests
    
    **Design Principles:**
    - Atomic operations are reusable across workflows
    - Business workflows orchestrate atomic operations
    - All code submissions use standardized formats
    - Async task pattern for long-running operations
  version: 1.0.0
  contact:
    name: LLT Assistant Team
servers:
  - url: http://localhost:8886/api/v1
    description: Local development server
  - url: https://llt-assistant.fly.dev/api/v1
    description: Production server
...
paths:
...
  /workflows/generate-tests:
    post:
      summary: "Feature 1: Submit generation request"
      description: |
        **Feature 1**
        
        Submit source code plus a short natural-language description to kick off
        the end-to-end test generation workflow. The backend performs scenario
        discovery and code generation asynchronously. The response only contains
        a `task_id` so the client can poll `/tasks/{task_id}` for progress and
        retrieve the generated tests when ready.
        
        **Input:** Python source code + ≤200 character description
        **Output:** Async task metadata
      tags:
        - Feature 1 - Test Generation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateTestsRequest'
      responses:
        '202':
          description: Generation started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenerateTestsResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '422':
          $ref: '#/components/responses/ValidationError'

# ============================================================================
# FEATURE 2: Coverage Optimization Workflow
# ============================================================================
...
  /tasks/{task_id}:
    get:
      summary: Get task status
      description: |
        Poll for async task status and results.
        Used for long-running operations like coverage optimization.
      tags:
        - Tasks
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Task found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskStatus'
        '404':
          description: Task not found

# ============================================================================
# COMPONENTS
# ============================================================================
components:
  schemas:
...
    CodeSubmission:
      type: object
      description: Standard format for submitting code to any endpoint
      required:
        - code
        - language
        - framework
      properties:
        code:
          type: string
          description: Source code content
        language:
          type: string
          enum: [python]
          default: python
        framework:
          type: string
          enum: [pytest, unittest]
          default: pytest
        metadata:
          type: object
          properties:
            file_path:
              type: string
            module_path:
              type: string
            git_context:
              type: object
              properties:
                commit_hash:
                  type: string
                branch:
                  type: string
...
    ClientMetadata:
      type: object
      description: VSCode extension metadata
      properties:
        extension_version:
          type: string
        vscode_version:
          type: string
        platform:
          type: string
        workspace_hash:
          type: string

    # ========================================================================
    # ATOMIC OPERATION REQUESTS/RESPONSES
    # ========================================================================
...
    GenerateTestsRequest:
      type: object
      required:
        - code_submission
        - user_description
      properties:
        code_submission:
          $ref: '#/components/schemas/CodeSubmission'
        user_description:
          type: string
          minLength: 1
          maxLength: 200
          description: Natural language description of the tests to generate
        config:
          type: object
          properties:
            auto_review_before_return:
              type: boolean
              default: true
            max_test_count:
              type: integer
              minimum: 1
              maximum: 20
              default: 8
            preferred_style:
              type: string
              enum: [pytest_parametrize, pytest_plain, unittest]
        client_metadata:
          $ref: '#/components/schemas/ClientMetadata'
    GenerateTestsResponse:
      type: object
      required:
        - task_id
        - status
      properties:
        task_id:
          type: string
          format: uuid
          description: Async task identifier for `/tasks/{task_id}`
        status:
          type: string
          enum: [pending, processing]
          description: Initial task status
        request_id:
          type: string
          description: Server-side identifier for tracing

    # ========================================================================
    # FEATURE 2: Coverage Optimization Schemas
    # ========================================================================
...
    TaskStatus:
      type: object
      required:
        - task_id
        - status
      properties:
        task_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending, processing, completed, failed]
        progress:
          type: number
          minimum: 0
          maximum: 1
        result:
          type: object
          description: Task result (when status=completed)
        error:
          type: object
          description: Error details (when status=failed)
          properties:
            message:
              type: string
            details:
              type: object
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        estimated_completion:
          type: string
          format: date-time

    # ========================================================================
    # SYSTEM SCHEMAS
    # ========================================================================
...
    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error type identifier
        message:
          type: string
          description: Human-readable error message
        details:
          type: object
          nullable: true
          additionalProperties: true
        request_id:
          type: string
          description: Request ID for debugging

  # ==========================================================================
  # REUSABLE RESPONSES
  # ==========================================================================
  responses:
    BadRequest:
      description: Malformed request or missing required fields
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: BadRequest
            message: Missing required field 'code' in request body
    ValidationError:
      description: Input failed domain validation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: ValidationError
            message: User description must be between 1 and 200 characters
            details:
              field: user_description
              constraint: maxLength(200)
...